✨ Add finesse...

+++ setup.sh
@@ -1,23 +1,29 @@
 #!/bin/bash
 
-# update stuff
-apt update
-apt upgrade -y
-apt autoremove -y
-apt autoclean -y
+echo
+echo "wassssuuuuup....!"
+echo "script will install basic tools and setupup configs for the new user."
+read $REPLY -p "sounds good? [y/n] "
+
+if [[ $REPLY =~ ^[Nn]$ ]]; then
+    echo "exiting..."
+    exit 1
+fi
 
 
-# Zsh
+# install stuff
+echo; echo "Installing packages..."
+apt install -y git tmux nodejs npm fonts-powerline
+npm -g install pnpm
+
+echo; echo "Conjuring Zsh magic..."
 apt install -y zsh
 chsh -s $(which zsh) # set zsh as default shell
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" # OhMyZsh
 
 
-# install stuff
-apt install -y zsh tmux git node fonts-powerline
-
-
 # config stuff
+echo; echo "Configuring stuff..."
 cat .zshrc >> ~/.zshrc
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 cp .tmux.conf ~/.tmux.conf
@@ -27,3 +33,7 @@ git config --global user.name "Ryan Norooz"
 git config --global user.email "58340667+RyanNorooz@users.noreply.github.com"
 git config --global user.username "RyanNorooz"
 git config --global core.editor "nano"
+
+
+echo; echo "Done!"
+echo "its a good idea to restart your system now..."
